Add a readme.md file to give details the work
This commit is dedicated to adding my readme.md file to my final project
because I now understand the importance of that file and how useful it
is to others. This project is an overview of the basics of HTML & CSS,
getting through the chapters will make anyone feel triumphant and this
code now serves as a resource of examples to come back to when I need a
refresh.